moved session class to separate file.
