+++ ys-client-server-1347/homework-1/ViewController.swift
@@ -7,16 +7,16 @@
 
 import UIKit
 
-class Session {
-    
-    static let instance = Session()
-    
-    private init() {}
-    
-    var userId: Int = 0
-    var token: String = ""
-    
-}
+//class Session {
+//    
+//    static let instance = Session()
+//
+//    private init() {}
+//
+//    var userId: Int = 0
+//    var token: String = ""
+//
+//}
 
 class ViewController: UIViewController {
