+++ ys-client-server-1347/homework-2/FirstViewController.swift
@@ -6,9 +6,12 @@
 //
 
 import UIKit
+import WebKit
 
 class FirstViewController: UIViewController {
-
+    
+    @IBOutlet weak var wk: WKWebView!
+    
     override func viewDidLoad() {
         super.viewDidLoad()
     }
